Reorganize adapter layer
