+++ 1-quickie-adapters/src/main/java/org/pinkcrazyunicorn/quickie/adapters/callbacks/profile/RemoveAvailableCallback.java
@@ -1,8 +1,8 @@
-package org.pinkcrazyunicorn.quickie.adapters.profile;
+package org.pinkcrazyunicorn.quickie.adapters.callbacks.profile;
 
 import org.pinkcrazyunicorn.quickie.application.profile.ProfileService;
 import org.pinkcrazyunicorn.quickie.domain.Food;
-import org.pinkcrazyunicorn.quickie.adapters.FoodMapper;
+import org.pinkcrazyunicorn.quickie.adapters.mappers.FoodMapper;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventAnswer;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventCallback;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventParameter;
@@ -11,13 +11,13 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
-public class AddAvailableCallback implements EventCallback {
+public class RemoveAvailableCallback implements EventCallback {
     private final ProfileService service;
     private final FoodMapper foodMapper;
 
-    public AddAvailableCallback(ProfileService service) {
+    public RemoveAvailableCallback(ProfileService service) {
         this.service = service;
-        this.foodMapper = new FoodMapper();
+        foodMapper = new FoodMapper();
     }
 
     @Override
@@ -26,9 +26,9 @@ public class AddAvailableCallback implements EventCallback {
         String foodString = data.get("food");
 
         Food food = this.foodMapper.fromString(foodString);
+        this.service.markUnavailable(name, food);
 
-        this.service.addToAvailable(name, food);
-        return new EventAnswer("Successfully added available food");
+        return new EventAnswer("Successfully removed opinion");
     }
 
     @Override
