+++ 1-quickie-adapters/src/main/java/org/pinkcrazyunicorn/quickie/adapters/callbacks/profile/RemoveOpinionCallback.java
@@ -1,8 +1,8 @@
-package org.pinkcrazyunicorn.quickie.adapters.profile;
+package org.pinkcrazyunicorn.quickie.adapters.callbacks.profile;
 
 import org.pinkcrazyunicorn.quickie.application.profile.ProfileService;
 import org.pinkcrazyunicorn.quickie.domain.Food;
-import org.pinkcrazyunicorn.quickie.adapters.FoodMapper;
+import org.pinkcrazyunicorn.quickie.adapters.mappers.FoodMapper;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventAnswer;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventCallback;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventParameter;
@@ -11,11 +11,11 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
-public class RemoveAvailableCallback implements EventCallback {
+public class RemoveOpinionCallback implements EventCallback {
     private final ProfileService service;
     private final FoodMapper foodMapper;
 
-    public RemoveAvailableCallback(ProfileService service) {
+    public RemoveOpinionCallback(ProfileService service) {
         this.service = service;
         foodMapper = new FoodMapper();
     }
@@ -26,7 +26,7 @@ public class RemoveAvailableCallback implements EventCallback {
         String foodString = data.get("food");
 
         Food food = this.foodMapper.fromString(foodString);
-        this.service.markUnavailable(name, food);
+        this.service.removeOpinionAbout(name, food);
 
         return new EventAnswer("Successfully removed opinion");
     }
