+++ 1-quickie-adapters/src/main/java/org/pinkcrazyunicorn/quickie/adapters/callbacks/recipe/ViewRecipesCallback.java
@@ -1,9 +1,10 @@
-package org.pinkcrazyunicorn.quickie.adapters.recipe;
+package org.pinkcrazyunicorn.quickie.adapters.callbacks.recipe;
 
 import org.pinkcrazyunicorn.quickie.adapters.event.EventAnswer;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventAnswerData;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventCallback;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventParameter;
+import org.pinkcrazyunicorn.quickie.adapters.mappers.recipe.RecipeMapper;
 import org.pinkcrazyunicorn.quickie.application.recipe.RecipeService;
 
 import java.util.Collection;
