+++ 1-quickie-adapters/src/main/java/org/pinkcrazyunicorn/quickie/adapters/mappers/profile/ProfileMapper.java
@@ -1,6 +1,6 @@
-package org.pinkcrazyunicorn.quickie.adapters.profile;
+package org.pinkcrazyunicorn.quickie.adapters.mappers.profile;
 
-import org.pinkcrazyunicorn.quickie.adapters.FoodMapper;
+import org.pinkcrazyunicorn.quickie.adapters.mappers.FoodMapper;
 import org.pinkcrazyunicorn.quickie.adapters.event.EventAnswerData;
 import org.pinkcrazyunicorn.quickie.adapters.event.ListEventAnswerData;
 import org.pinkcrazyunicorn.quickie.adapters.event.MapEventAnswerData;
